#118 - groupMemberService 로 팀 멤버 관련 로직 분리
