+++ src/main/java/com/ting/ting/configuration/AppConfig.java
@@ -28,6 +28,11 @@ public class AppConfig {
     }
 
     @Bean
+    public GroupMemberService groupMemberService() {
+        return new GroupMemberServiceImpl(userRepository, groupRepository, groupMemberRepository, groupMemberRequestRepository);
+    }
+
+    @Bean
     public GroupLikeService groupLikeService() {
         return new GroupLikeServiceImpl(userRepository, groupRepository, groupMemberRepository, groupMemberRequestRepository, groupDateRequestRepository, groupLikeToJoinRepository, groupLikeToDateRepository);
     }
