+++ src/main/java/com/ting/ting/controller/GroupControllerImpl.java
@@ -4,6 +4,7 @@ import com.ting.ting.dto.request.GroupRequest;
 import com.ting.ting.dto.response.*;
 import com.ting.ting.exception.ServiceType;
 import com.ting.ting.service.GroupLikeService;
+import com.ting.ting.service.GroupMemberService;
 import com.ting.ting.service.GroupService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -16,11 +17,13 @@ import java.util.Set;
 public class GroupControllerImpl extends AbstractController implements GroupController {
 
     private final GroupService groupService;
+    private final GroupMemberService groupMemberService;
     private final GroupLikeService groupLikeService;
 
-    public GroupControllerImpl(GroupService groupService, GroupLikeService groupLikeService) {
+    public GroupControllerImpl(GroupService groupService, GroupMemberService groupMemberService, GroupLikeService groupLikeService) {
         super(ServiceType.GROUP_MEETING);
         this.groupService = groupService;
+        this.groupMemberService = groupMemberService;
         this.groupLikeService = groupLikeService;
     }
 
@@ -92,67 +95,9 @@ public class GroupControllerImpl extends AbstractController implements GroupCont
     }
 
     @Override
-    public Response<Set<GroupMemberResponse>> getGroupMemberList(Long groupId) {
-        return success(groupService.findGroupMemberList(groupId));
-    }
-
-    @Override
     public Response<GroupResponse> createGroup(GroupRequest request) {
         Long userId = 9L;  // userId를 임의로 설정 TODO: user 구현 후 수정
         return success(groupService.saveGroup(userId, request));
-    }
-
-    @Override
-    public Response<Void> sendJoinRequest(Long groupId) {
-        Long userId = groupId + 1;  // userId를 임의로 설정 TODO: user 구현 후 수정
-
-        groupService.saveJoinRequest(groupId, userId);
-        return success();
-    }
-
-    @Override
-    public Response<Void> deleteJoinRequest(Long groupId) {
-        Long userId = groupId + 1;  // userId를 임의로 설정 TODO: user 구현 후 수정
-
-        groupService.deleteJoinRequest(groupId, userId);
-        return success();
-    }
-    
-    @Override
-    public Response<Void> deleteGroupMember(Long groupId) {
-        Long userId = 1L;
-
-        groupService.deleteGroupMember(groupId, userId);
-        return success();
-    }
-
-    @Override
-    public Response<Set<GroupMemberResponse>> changeGroupLeader(Long groupId, Long userIdOfNewLeader) {
-        Long userIdOfLeader = 1L;  // userId를 임의로 설정 TODO: user 구현 후 수정
-
-        return success(groupService.changeGroupLeader(groupId, userIdOfLeader, userIdOfNewLeader));
-    }
-
-    @Override
-    public Response<Set<GroupMemberRequestResponse>> getMemberRequestToJoinMyGroup(Long groupId) {
-        Long userIdOfLeader = 1L;  // userId를 임의로 설정 TODO: user 구현 후 수정
-
-        return success(groupService.findMemberJoinRequest(groupId, userIdOfLeader));
-    }
-
-    @Override
-    public Response<GroupMemberResponse> acceptJoinRequestToMyGroup(Long groupMemberRequestId) {
-        Long userIdOfLeader = 1L;  // userId를 임의로 설정 TODO: user 구현 후 수정
-
-        return success(groupService.acceptMemberJoinRequest(userIdOfLeader, groupMemberRequestId));
-    }
-
-    @Override
-    public Response<Void> rejectJoinRequestToMyGroup(Long groupMemberRequestId) {
-        Long userIdOfLeader = 1L; // userId를 임의로 설정 TODO: user 구현 후 수정
-
-        groupService.rejectMemberJoinRequest(userIdOfLeader, groupMemberRequestId);
-        return success();
     }
 
     @Override
@@ -189,6 +134,64 @@ public class GroupControllerImpl extends AbstractController implements GroupCont
         Long userIdOfLeader = 1L; // userId를 임의로 설정 TODO: user 구현 후 수정
 
         groupService.rejectGroupDateRequest(userIdOfLeader, groupDateRequestId);
+        return success();
+    }
+
+    @Override
+    public Response<Set<GroupMemberResponse>> getGroupMemberList(Long groupId) {
+        return success(groupMemberService.findGroupMemberList(groupId));
+    }
+
+    @Override
+    public Response<Void> sendJoinRequest(Long groupId) {
+        Long userId = groupId + 1;  // userId를 임의로 설정 TODO: user 구현 후 수정
+
+        groupMemberService.saveJoinRequest(groupId, userId);
+        return success();
+    }
+
+    @Override
+    public Response<Void> deleteJoinRequest(Long groupId) {
+        Long userId = groupId + 1;  // userId를 임의로 설정 TODO: user 구현 후 수정
+
+        groupMemberService.deleteJoinRequest(groupId, userId);
+        return success();
+    }
+
+    @Override
+    public Response<Void> deleteGroupMember(Long groupId) {
+        Long userId = 1L;
+
+        groupMemberService.deleteGroupMember(groupId, userId);
+        return success();
+    }
+
+    @Override
+    public Response<Set<GroupMemberResponse>> changeGroupLeader(Long groupId, Long userIdOfNewLeader) {
+        Long userIdOfLeader = 1L;  // userId를 임의로 설정 TODO: user 구현 후 수정
+
+        return success(groupMemberService.changeGroupLeader(groupId, userIdOfLeader, userIdOfNewLeader));
+    }
+
+    @Override
+    public Response<Set<GroupMemberRequestResponse>> getMemberRequestToJoinMyGroup(Long groupId) {
+        Long userIdOfLeader = 1L;  // userId를 임의로 설정 TODO: user 구현 후 수정
+
+        return success(groupMemberService.findMemberJoinRequest(groupId, userIdOfLeader));
+    }
+
+    @Override
+    public Response<GroupMemberResponse> acceptJoinRequestToMyGroup(Long groupMemberRequestId) {
+        Long userIdOfLeader = 1L;  // userId를 임의로 설정 TODO: user 구현 후 수정
+
+        return success(groupMemberService.acceptMemberJoinRequest(userIdOfLeader, groupMemberRequestId));
+    }
+
+    @Override
+    public Response<Void> rejectJoinRequestToMyGroup(Long groupMemberRequestId) {
+        Long userIdOfLeader = 1L; // userId를 임의로 설정 TODO: user 구현 후 수정
+
+        groupMemberService.rejectMemberJoinRequest(userIdOfLeader, groupMemberRequestId);
         return success();
     }
 }
